chore: reorg code with split
