+++ data_processing/dvf/geoloc/utils/latlong.py
@@ -1,0 +1,156 @@
+import logging
+import gc
+import pandas as pd
+import pyarrow.fs as pafs
+import pyarrow.parquet as pq
+
+from datagouvfr_data_pipelines.utils.s3 import S3Client
+
+
+def match_year_to_snapshots(year, available_dates) -> list[str]:
+    """Match year to available snapshots dates.
+    Returns a list of ISO str dates ["YYYY-MM-DD", ...]"""
+    restr_available_dates = [
+        max(
+            k for k in available_dates.keys() if k.startswith(f"{int(year) - 1}")
+        )  # take last snapshot of the prior year, e.g. "2022-10-01"
+    ] + sorted(
+        [k for k in available_dates.keys() if k.startswith(year)]
+    )  # all snapshots of the current year
+    if restr_available_dates[-1] < f"{year}-12-31":
+        restr_available_dates.append(
+            f"{year}-12-31"
+        )  # closing bound of the current year, not a snapshot
+    return restr_available_dates
+
+
+def enrich_parcelles_with_snapshot_coord(
+    dvf_batch_df: pd.DataFrame,
+    snapshot_file: str,
+    s3_client: S3Client,
+    bucket: str,
+) -> pd.DataFrame:
+    """Match parcelle's lat, long coordinates for a snapshot file to the DVF batch,
+    relying on parcelle ID.
+    Note: snapshot parquet files are produced by Thomas G.
+    """
+    logging.info("Streaming " + snapshot_file)
+    parcelle_ids = pd.Index(dvf_batch_df["id_parcelle"].unique())
+
+    # Open just once the file to accelerate time
+    fs = pafs.S3FileSystem(  # type: ignore - lazily init. by pafs GH-38364
+        access_key=s3_client.login,
+        secret_key=s3_client.password,
+        endpoint_override=s3_client.url,
+        scheme="https",
+    )
+    coord_pf = pq.ParquetFile(
+        f"{bucket}/{snapshot_file}", filesystem=fs, pre_buffer=True
+    )
+    logging.info(
+        f"> {coord_pf.metadata.num_rows:,} parcelles in {coord_pf.metadata.num_row_groups} row groups"
+    )
+    coord_to_keep_chunks = []
+    for i, coord_batch in enumerate(
+        coord_pf.iter_batches(
+            batch_size=100_000, columns=["id", "latitude", "longitude"]
+        )
+    ):
+        coord_batch_df = coord_batch.to_pandas()
+        coord_to_keep = coord_batch_df.loc[coord_batch_df["id"].isin(parcelle_ids)]
+        if not coord_to_keep.empty:
+            coord_to_keep_chunks.append(coord_to_keep)
+        del coord_batch_df, coord_batch, coord_to_keep
+        if i % 20 == 0:
+            logging.info(
+                f"> batch {i}, {sum(map(len, coord_to_keep_chunks)):,} parcelles matched so far"
+            )
+    del parcelle_ids
+    if not coord_to_keep_chunks:
+        raise ValueError(
+            f"No matching parcelles ID between DVF batch and coordinates snapshot {snapshot_file}"
+        )
+    coord_to_keep_df = pd.concat(coord_to_keep_chunks, ignore_index=True).rename(
+        {"id": "id_parcelle"}, axis=1
+    )
+    assert (
+        coord_to_keep_df["id_parcelle"].is_unique
+    )  # if id is NOT unique in the parquet file, duplicates would appear in the join
+    del coord_to_keep_chunks
+    gc.collect()
+
+    enriched_batch_df = pd.merge(
+        dvf_batch_df, coord_to_keep_df, on="id_parcelle", how="left"
+    )
+    del coord_to_keep_df
+    logging.info(
+        f"> {round(len(enriched_batch_df.loc[enriched_batch_df['latitude'].isna()]) / len(enriched_batch_df) * 100, 2)}% missing"
+    )  # expecting around 1-2% missing coords
+    return enriched_batch_df
+
+
+def enrich_parcelles_with_coord(
+    dvf_df: pd.DataFrame,
+    year: str,
+    available_dates: dict[str, str],
+    s3_client: S3Client,
+    bucket: str,
+):
+    """Enrich DVF dataframe with parcelle's lat, long coordinates.
+    Those parcelle coordinates can be defined at a slightly different timeframe than the mutation,
+    so looping through each available snapshot of parcelle coordinates.
+    """
+    restr_available_dates = match_year_to_snapshots(year, available_dates)
+    logging.info(restr_available_dates)
+    geoloced = []
+    remainders = None
+    for k in range(len(restr_available_dates) - 1):
+        dmin, dmax = (
+            restr_available_dates[k],
+            restr_available_dates[k + 1],
+        )  # lower and upper date bounds for current snapshot
+        matching_mutations_df = dvf_df.loc[
+            dvf_df["date_mutation"].between(
+                dmin, dmax, inclusive="both" if dmax == f"{year}-12-31" else "left"
+            )
+        ]  # the rows between the snapshot date bounds to process
+        dvf_df.drop(
+            matching_mutations_df.index, inplace=True
+        )  # dropping them from original df to keep RAM low
+        logging.info(f"{len(matching_mutations_df)} rows between {dmin} and {dmax}")
+        if remainders is not None and not remainders.empty:
+            # for parcelles that didn't get geolocalized in the expected batch, we'll try again at each upcoming batch
+            logging.info(f"- adding {len(remainders)} remainders")
+            matching_mutations_df = pd.concat(
+                [matching_mutations_df, remainders], ignore_index=True
+            )
+        if len(matching_mutations_df) == 0:
+            logging.info("> skipping")
+            continue
+        # TODO: sorting below preserve row order from prior implem. Need to verify if necessary otherwise can be removed
+        matching_mutations_df = matching_mutations_df.sort_values(
+            by="id_parcelle", key=lambda s: s.str[:3], kind="stable", ignore_index=True
+        )  # sort by 2 digits department + 1st digit of the commune code like "380"
+        enriched = enrich_parcelles_with_snapshot_coord(
+            matching_mutations_df, available_dates[dmin], s3_client, bucket
+        )
+        remainders = enriched.loc[enriched["longitude"].isna()][
+            [c for c in enriched.columns if c not in ["latitude", "longitude"]]
+        ]
+        geoloced.append(enriched.dropna(subset="longitude"))
+        del enriched
+        gc.collect()
+    logging.info("Done with geoloc, concatenating results...")
+    if remainders is not None:
+        geoloced.append(remainders)
+    del remainders
+
+    # using pd.concat on geoloced directly is too RAM heavy, so workaround
+    enriched_df = pd.DataFrame()
+    while geoloced:
+        logging.info(f"> {len(geoloced)} dfs still to concatenate")
+        enriched_df = pd.concat([enriched_df, geoloced[0]], ignore_index=True)
+        del geoloced[0]
+        gc.collect()
+    del geoloced
+    return enriched_df
